Added bookmarks feature.
A bookmark button was added that presents directories that were
previously saved as bookmarks. The corresponding table view is
presented using PanelKit.

A + Button allows to add a bookmark for the current directory. Upon
selecting a bookmark in the table view, the current directory is
changed to the bookmarked directory. A saved bookmark can be deleted by
swiping the corresponding table view cell.

In order to keep permissions to access directories outside of the
sandbox, bookmarks are stored as bookmark data in ~/.bookmarks.