Fixed an issue with directories with the same name.
Until now, if you had two directories called e.g. "Documents" in
different paths, saving a bookmark to one of them would overwrite the
other, as only the last path component was used to determine the
bookmark file name. Now the directory path is used to create the file
name, resolving this issue.

Additionally, the "file://" part of the URL is no longer displayed in
the bookmarks list, making it look a little cleaner.