Added bookmarks feature. (#65)
* Added bookmarks feature.

A bookmark button was added that presents directories that were
previously saved as bookmarks. The corresponding table view is
presented using PanelKit.

A + Button allows to add a bookmark for the current directory. Upon
selecting a bookmark in the table view, the current directory is
changed to the bookmarked directory. A saved bookmark can be deleted by
swiping the corresponding table view cell.

In order to keep permissions to access directories outside of the
sandbox, bookmarks are stored as bookmark data in ~/.bookmarks.

* Moved bookmark managing code from ViewController to BookmarkManager.

In order to not clutter the ViewController class, bookmark managing was
moved to the BookmarkManager class.

* Removed the duplicate panel background color.

* The directory to store the bookmarks is now obtained using the DocumentManager.

* Single line comments only use one line now.

* Refactored didSelectBookmarkURL to changeDirectoryToURL.

As the documentPicker and the bookmark selection did the very same
thing, the didSelectBookmarkURL is now changeDirectoryToURL, which the
documentPicker uses.

* Removed an unnecessary re-throw.

* When loading a bookmark fails, the corresponding file is now removed.

Also, the loop will no longer break when a single url couldn’t be
loaded.

* The bookmark view controller now displays sanitized URLs.

* Removed an unnecessary delegate requirement.

* Fixed an issue with directories with the same name.

Until now, if you had two directories called e.g. "Documents" in
different paths, saving a bookmark to one of them would overwrite the
other, as only the last path component was used to determine the
bookmark file name. Now the directory path is used to create the file
name, resolving this issue.

Additionally, the "file://" part of the URL is no longer displayed in
the bookmarks list, making it look a little cleaner.